% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coanc_admix.R
\name{coanc-deprecated}
\alias{coanc-deprecated}
\title{Construct the coancestry matrix of an admixture model}
\usage{
coanc(Q, F)
}
\arguments{
\item{Q}{The admixture proportion matrix}

\item{F}{Either the intermediate subpopulation coancestry matrix, or the vector of intermediate subpopulation \eqn{F_{ST}}{FST} values, or a scalar \eqn{F_{ST}}{FST} value shared by all intermediate subpopulations.}
}
\value{
The coancestry matrix
}
\description{
Construct the coancestry matrix of an admixture model
}
\seealso{
\code{\link{bnpsd-deprecated}}
}
\keyword{internal}
