% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coanc_tree.R
\name{coanc_tree}
\alias{coanc_tree}
\title{Calculate coancestry matrix corresponding to a tree}
\usage{
coanc_tree(tree)
}
\arguments{
\item{tree}{The coancestry tree relating the \code{k_subpops} subpopulations.
Must be a \code{phylo} object from the \code{ape} package (see \code{\link[ape:read.tree]{ape::read.tree()}}).}
}
\value{
The \code{k_subpops}-by-\code{k_subpops} coancestry matrix.
}
\description{
This function calculates the coancestry matrix of the subpopulations which are tip nodes in the input tree.
The edges of this tree are coancestry values relative to the parent node of each child node.
}
\details{
The calculation takes into account that total coancestries are non-linear combinations of the per-edge coancestries.
For example, if the root node is \code{A}, and subpopulation \code{C} is connected to \code{A} only through an internal node \code{B}, then its total self-coancestry \code{coanc_A_C} relative to \code{A} is given by \code{coanc_A_B} (the coancestry between \code{A} and \code{B}) and \code{coanc_B_C} (the coancestry between \code{B} and \code{C}) by
\code{coanc_A_C = coanc_A_B + coanc_B_C * ( 1 - coanc_A_B )}.
This transformation ensures that the total coancestry is a probability that does not exceed one even when the per-edge coancestries sum to a value greater than one.
This transformation is iterated for all nodes in the tree.
The self-coancestry (diagonal values) are the total coancestries of the tip nodes.
The coancestry between different subpopulations is the total coancestry of their last common ancestor node.
}
\examples{
# for simulating a tree with `rtree`
library(ape)

# a typical, non-trivial example
# number of tip subpopulations
k_subpops <- 3
# simulate a random tree
tree_subpops <- rtree( k_subpops )
# coancestry matrix of subpopulations
coancestry <- coanc_tree( tree_subpops )

}
