% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_all_unstructured.R
\name{draw_all_unstructured}
\alias{draw_all_unstructured}
\title{Simulate random allele frequencies and genotypes from the unstructured model}
\usage{
draw_all_unstructured(
  n_ind,
  m_loci = NA,
  beta = NA,
  p_anc = NULL,
  require_polymorphic_loci = TRUE,
  maf_min = 0,
  verbose = TRUE
)
}
\arguments{
\item{n_ind}{The number of individuals to draw (required).}

\item{m_loci}{The number of loci to draw.  Required except when \code{p_anc} below is provided and is a vector, in which case the number of loci equals the length of \code{p_anc} and the value of \code{m_loci} passed is ignored.}

\item{beta}{Shape parameter for a symmetric Beta for ancestral allele frequencies \code{p_anc}.
If \code{NA} (default), \code{p_anc} is uniform with range in [0.01, 0.5].
Otherwise, \code{p_anc} has a symmetric Beta distribution with range in [0, 1].
Has no effect if \code{p_anc} option is non-\code{NULL}.}

\item{p_anc}{If provided, it is used as the ancestral allele frequencies (instead of drawing random ones).  Must either be a scalar or a length-\code{m_loci} vector.
If scalar, \code{m_loci} is required, and the returned \code{p_anc} is the scalar value repeated \code{m_loci} times.
If \code{p_anc} is a vector, its length is used to define \code{m_loci} and the value of \code{m_loci} passed is ignored.
If a locus was fixed and has to be redrawn, the ancestral allele frequency in \code{p_anc} is retained and only genotypes are redrawn.}

\item{require_polymorphic_loci}{If \code{TRUE} (default), returned genotype matrix will not include any fixed loci (loci that happened to be fixed are drawn again, starting from their ancestral allele frequencies, and checked iteratively until no fixed loci remain, so that the final number of polymorphic loci is exactly \code{m_loci}).}

\item{maf_min}{The minimum minor allele frequency (default zero), to extend the working definition of "fixed" above to include rare variants.
This helps simulate a frequency-based locus ascertainment bias.
Loci with minor allele frequencies less than or equal to this value are treated as fixed (passed to \code{\link[=fixed_loci]{fixed_loci()}}).
This parameter has no effect if \code{require_polymorphic_loci} is \code{FALSE}.}

\item{verbose}{If \code{TRUE}, prints messages for every stage in the algorithm.}
}
\value{
A named list with the following items (which may be missing depending on options):
\itemize{
\item \code{X}: An \code{m_loci}-by-\code{n_ind} matrix of genotypes.
\item \code{p_anc}: A length-\code{m_loci} vector of ancestral allele frequencies.
}
}
\description{
This function returns simulated ancestral allele frequencies and genotypes without structure, meaning individuals draw their genotypes independently and identically from the Binomial distribution with the same ancestral allele frequency per locus.
The function is a wrapper around \code{\link[=draw_p_anc]{draw_p_anc()}} with additional features such as requiring polymorphic loci, mimicking \code{\link[=draw_all_admix]{draw_all_admix()}} in options as applicable.
Importantly, by default fixed loci (where all individuals were homozygous for the same allele) are re-drawn from the start (starting from the ancestral allele frequencies) so no fixed loci are in the output.
Below \code{m_loci} (also \code{m}) is the number of loci and \code{n_ind} is the number of individuals.
}
\examples{
# dimensions
# number of loci
m_loci <- 10
# number of individuals
n_ind <- 5

# draw all random allele freqs and genotypes
out <- draw_all_unstructured( n_ind, m_loci )

# return value is a list with these items:

# genotypes
X <- out$X

# ancestral AFs
p_anc <- out$p_anc

}
