% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_prop_1d_circular.R
\name{q1dc-deprecated}
\alias{q1dc-deprecated}
\title{Construct admixture proportion matrix for 1D geography}
\usage{
q1dc(n, k, sigma, a = 0, b = 2 * pi, s, F, Fst)
}
\arguments{
\item{n}{Number of individuals}

\item{k}{Number of intermediate subpopulations}

\item{sigma}{Spread of intermediate subpopulations (approximate standard deviation of Von Mises densities, see above)}

\item{a}{Location of first individual}

\item{b}{Location of last individual

OPTIONS FOR BIAS COEFFICIENT VERSION}

\item{s}{The desired bias coefficient, which specifies \eqn{\sigma} indirectly.  Required if \code{sigma} is missing}

\item{F}{The vector of inbreeding coefficients of the intermediate subpopulations, up to a scaling factor (which cancels out in calculations).  Required if \code{sigma} is missing}

\item{Fst}{The desired final \eqn{F_{ST}}{FST} of the admixed individuals.  Required if \code{sigma} is missing}
}
\value{
If \code{sigma} was provided, the \eqn{n \times k}{n-by-k} admixture proportion matrix \eqn{Q}.
If \code{sigma} is missing, a named list is returned containing \code{Q}, the rescaled \code{F}, and the \code{sigma} that together give the desired \eqn{s} and final \eqn{F_{ST}}{FST} of the admixed individuals.
}
\description{
Construct admixture proportion matrix for 1D geography
}
\seealso{
\code{\link{bnpsd-deprecated}}
}
\keyword{internal}
