% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_prop_indep_subpops.R
\name{qis-deprecated}
\alias{qis-deprecated}
\title{Construct admixture proportion matrix for independent subpopulations}
\usage{
qis(labs, subpops)
}
\arguments{
\item{labs}{Vector of subpopulation labels.}

\item{subpops}{Optional vector of unique subpopulations in desired order.}
}
\value{
The admixture proportion matrix.
}
\description{
Construct admixture proportion matrix for independent subpopulations
}
\seealso{
\code{\link{bnpsd-deprecated}}
}
\keyword{internal}
