% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_genotypes_admix.R
\name{rgeno-deprecated}
\alias{rgeno-deprecated}
\title{Draw genotypes from the admixture model}
\usage{
rgeno(P, Q = NULL, lowMem = FALSE)
}
\arguments{
\item{P}{The IAF matrix (if Q is missing) or the intermediate subpopulation allele frequency matrix (if Q is present)}

\item{Q}{The optional admixture proportion matrix}

\item{lowMem}{OBSOLETE (the default algorithm is now the old lowMem=TRUE)}
}
\value{
The genotype matrix
}
\description{
Draw genotypes from the admixture model
}
\seealso{
\code{\link{bnpsd-deprecated}}
}
\keyword{internal}
