% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_p_anc.R
\name{rpanc-deprecated}
\alias{rpanc-deprecated}
\title{Draw uniform ancestral allele frequencies}
\usage{
rpanc(m, min = 0.01, max = 0.5)
}
\arguments{
\item{m}{Number of loci}

\item{min}{Minimum allele frequency to draw}

\item{max}{Maximum allele frequency to draw}
}
\value{
A vector of ancestral allele frequencies
}
\description{
Draw uniform ancestral allele frequencies
}
\seealso{
\code{\link{bnpsd-deprecated}}
}
\keyword{internal}
