% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpanc.R
\name{rpanc}
\alias{rpanc}
\title{Draw uniform ancestral allele frequencies}
\usage{
rpanc(m, min = 0.01, max = 0.5)
}
\arguments{
\item{m}{Number of loci}

\item{min}{Minimum allele frequency to draw}

\item{max}{Maximum allele frequency to draw}
}
\value{
A length-\eqn{m} vector of ancestral allele frequencies
}
\description{
This is simply a wrapper around \code{\link[stats]{runif}} with different defaults and additional validations.
}
\examples{
pAnc <- rpanc(m = 10)

}
