% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_p_ind_admix.R
\name{rpiaf-deprecated}
\alias{rpiaf-deprecated}
\title{Construct individual-specific allele frequency matrix under the PSD admixture model}
\usage{
rpiaf(B, Q)
}
\arguments{
\item{B}{The matrix of intermediate subpopulation allele frequencies.}

\item{Q}{The matrix of admixture proportions.}
}
\value{
The matrix of individual-specific allele frequencies.
}
\description{
Construct individual-specific allele frequency matrix under the PSD admixture model
}
\seealso{
\code{\link{bnpsd-deprecated}}
}
\keyword{internal}
