% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_p_subpops.R
\name{rpint-deprecated}
\alias{rpint-deprecated}
\title{Draw intermediate subpopulation allele frequencies}
\usage{
rpint(pAnc, F)
}
\arguments{
\item{pAnc}{The vector of ancestral allele frequencies per locus}

\item{F}{The vector of subpopulation \eqn{F_{ST}}{FST} values}
}
\value{
The matrix of intermediate subpopulation allele frequencies
}
\description{
Draw intermediate subpopulation allele frequencies
}
\seealso{
\code{\link{bnpsd-deprecated}}
}
\keyword{internal}
