% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_p_anc.R
\name{draw_p_anc}
\alias{draw_p_anc}
\title{Draw random uniform ancestral allele frequencies}
\usage{
draw_p_anc(m_loci, p_min = 0.01, p_max = 0.5)
}
\arguments{
\item{m_loci}{Number of loci to draw}

\item{p_min}{Minimum allele frequency to draw}

\item{p_max}{Maximum allele frequency to draw}
}
\value{
A length-\eqn{m} vector of random ancestral allele frequencies
}
\description{
This is simply a wrapper around \code{\link[stats]{runif}} with different defaults and additional validations.
}
\examples{
p_anc <- draw_p_anc(m_loci = 10)

}
