% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_p_subpops.R
\name{draw_p_subpops}
\alias{draw_p_subpops}
\title{Draw allele frequencies for independent subpopulations}
\usage{
draw_p_subpops(p_anc, inbr_subpops)
}
\arguments{
\item{p_anc}{The length-\eqn{m} vector of ancestral allele frequencies per locus}

\item{inbr_subpops}{The length-\eqn{k} vector of subpopulation \eqn{F_{ST}}{FST} values}
}
\value{
The \eqn{m \times k}{m-by-k} matrix of independent subpopulation allele frequencies
}
\description{
Allele frequencies \eqn{p_i^{S_u}} for independent subpopulations \eqn{S_u} at locus \eqn{i} are drawn from the Balding-Nichols distribution with ancestral allele frequency \eqn{p_i^T} and \eqn{F_{ST}}{FST} parameter \eqn{f^T_{S_u}} as
\deqn{p_i^{S_u} \sim \mbox{Beta}(\nu_u p_i^T, \nu_u (1-p_i^T)),}
where \eqn{\nu_u = 1/f^T_{S_u} - 1}.
Below \eqn{m} is the number of loci and \eqn{k} is the number of subpopulations.
}
\examples{
m_loci <- 10 # number of loci
p_anc <- draw_p_anc(m_loci) # random vector of ancestral allele frequencies
inbr_subpops <- c(0.1, 0.3) # FST values for two subpops
p_subpops <- draw_p_subpops(p_anc, inbr_subpops) # matrix of intermediate subpop allele freqs

}
